requirements.txt
torch
torchvision
torchaudio
timm==0.6.13
matplotlib==3.7.1
numpy==1.24.0
scikit-learn==1.2.2
Pillow==9.5.0
pandas==1.5.3
tqdm==4.65.0
seaborn
streamlit
